refactor: Rename extension modules.
