+++ src/extensions/compress.rs
