+++ src/extensions/idle.rs
