+++ src/imap4rev1.rs
@@ -1,7 +1,7 @@
 use imap_types::message::AuthMechanism;
 use nom::IResult;
 
-use crate::rfc3501::core::atom;
+use crate::imap4rev1::core::atom;
 
 pub mod address;
 pub mod body;
