+++ src/imap4rev1/envelope.rs
@@ -12,7 +12,7 @@ use nom::{
     IResult,
 };
 
-use crate::rfc3501::{
+use crate::imap4rev1::{
     address::address,
     core::{nil, nstring},
 };
