+++ src/imap4rev1/flag.rs
@@ -10,7 +10,7 @@ use nom::{
     IResult,
 };
 
-use crate::rfc3501::core::atom;
+use crate::imap4rev1::core::atom;
 
 /// ```abnf
 /// flag = "\Answered" /
